# internal/apidocs/openapi.yaml
openapi: 3.1.0
info:
  title: Bekci API
  version: "1"
  description: |
    Bekci'nin (kendi sunucunuzda çalışan uptime ve sunucu izleme) REST API'si.
    Arayüzün kullandığı API ile aynıdır; betikler ve Terraform/Ansible gibi
    araçlar **API anahtarı** ile kullanır.

    **Kimlik doğrulama:** `Authorization: Bearer upk_…` (Ayarlar → Hesabım →
    API anahtarları). Tarayıcı oturumu (çerez) ile çağrılan yazma istekleri
    `X-Uptime: 1` başlığı ister (CSRF); API anahtarıyla gerek yoktur.
    Hesap güvenliği uçları (`/api/auth/*`, kullanıcı yönetimi) API anahtarıyla
    kullanılamaz.

    **Roller:** izleyici (viewer) okur; editör (editor) monitör, kanal, durum
    sayfası ve bakım yönetir; yönetici (admin) her şeyi. Yanıt dili
    `X-Uptime-Lang: tr|en` ya da `Accept-Language` ile seçilir.

    **Zamanlar** Unix saniyedir (`started_at`, `created_at` …); `0` "yok" demektir.
    Hatalar `{"error": "…", "code": "…"}` gövdesiyle döner (code isteğe bağlı).
  license:
    name: AGPL-3.0
    url: https://github.com/kadirsungurlu/bekci/blob/main/LICENSE
servers:
  - url: /
security:
  - apiKey: []
  - session: []
...
components:
...
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: API anahtarı (`upk_…`); rolü anahtarın rolü ile sahibinin rolünün küçüğüdür.
    session:
      type: apiKey
      in: cookie
      name: uptime_oturum
      description: "Tarayıcı oturumu; yazma isteklerinde `X-Uptime: 1` başlığı gerekir."
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    uid:
      name: uid
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
    Error:
      description: Hata
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    OK:
      description: Başarılı
      content:
        application/json:
          schema: { $ref: "#/components/schemas/OK" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        code: { type: string, description: "İsteğe bağlı makine kodu (ör. password_change_required, backup_password, reset_invalid)" }
      required: [error]
    OK:
      type: object
      properties:
        ok: { type: boolean }
...
    IncidentUpdate:
      type: object
      properties:
        id: { type: integer, format: int64 }
        incident_id: { type: integer, format: int64 }
        time: { type: integer, format: int64 }
        state: { type: string }
        body: { type: string }
        user_id: { type: integer, format: int64 }
        username: { type: string }
...
paths:
...
  /api/incidents/{id}/updates:
    post:
      tags: [incidents]
      summary: Güncelleme yaz (aşama + metin; editör+)
      parameters: [{ $ref: "#/components/parameters/id" }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { state: { type: string, enum: [investigating, identified, monitoring, resolved] }, body: { type: string } } } } } }
      responses: { "201": { description: Güncelleme, content: { application/json: { schema: { $ref: "#/components/schemas/IncidentUpdate" } } } } }
  /api/incidents/{id}/updates/{uid}:
    delete: { tags: [incidents], summary: Güncellemeyi sil (editör+), parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/uid" }], responses: { "200": { $ref: "#/components/responses/OK" } } }
...
  /api/probes/{id}/update:
    post: { tags: [servers], summary: "Ajanı panelin sürümüne şimdi güncelle (yönetici; otomatik güncelleme kapalı olsa da bir kez)", parameters: [{ $ref: "#/components/parameters/id" }], responses: { "200": { description: Kontrol noktası / ajan }, "400": { $ref: "#/components/responses/Error" } } }
  /api/probes/update-all:
    post:
      tags: [servers]
      summary: Verilen türdeki güncellenebilir tüm ajanlar için güncelleme iste (yönetici)
      parameters: [{ name: kind, in: query, required: true, schema: { type: string, enum: [server, location] } }]
      responses: { "200": { description: "{requested: N}" } }
...